+++ Source/RetSpoof/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(OsirisOBJ PRIVATE
+  RetSpoofGadgets.cpp
+)
