+++ Source/RetSpoof/RetSpoofGadgets.cpp
